Plugin filter/search and channel colors
Added plugin search feature: The remote plugin popup menu has been
replaced by a search window for improved access to the plugin library.

Added an option to disable server side filtering of plugins: By default
the server is matching the channel config of a client to the number of
inputs and outputs of the available plugins and returns only the
matching ones. As some plugins are not reporting their correct
input/output configuration, the server might not return it them to the
client.

Added a visualization of the channel color within the AG plugin UI for
better orientation (works only in VST3 right now).