Server: Move sandbox deletion to separate threads
This is to prevent hanging threads on Windows where the interprocess
comm passes a very long timeout to the pipe I/O APIs which might block
the thread deleting a peer when waiting for an I/O event. This area
needs improvements! The interprocess comm will likely have to be
rewritten to avoid JUCEs implementation.