Plugin: Added realtime/live mode and more stats to the monitor
RealTime mode

When this mode is enabled, the plugin sets strict timeouts in the audio
streamer that must me met for realtime processing. The plugin will drop
samples if not and tries to unblock the DAW that way.

Plugin monitor stats

Read queue level stats (avg, 95th) have been added as well as the number
of read errors. The performance statistic is now based on the most
recent average value rather then the 95th 1min value.